requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-registration
django-toolbelt==0.0.1
django-postgrespool==0.3.0
gunicorn==19.3.0
whitenoise==1.0.6
macros
pymongo
django-bootstrap-toolkit
nose
dj-static==0.0.6
static==0.4
wsgiref==0.1.2
Pillow==3.0
pyyaml
bcrypt
Fabric
unidecode
django-easy-maps
